Preparations for MS TTS
Create a new playURL function that only plays a sound, and make
xxxxSay() functions call it. It will allow playURL to be invoked
from an asynchronous callback function, as the ones required by
Microsoft TTS.

Still working in a solution to allow requesting MS keys without
restoring to GM_xmlhttpRequest.